Issue #132 Reuse setup code in parler_po tests
